requirements.txt
matplotlib==3.10.3
numpy==2.2.5
opencv_python==4.11.0.86
pandas==2.2.3
seaborn==0.13.2
torch==2.7.0
torchvision==0.22.0
ttach==0.0.3
YOLOv8_Explainer==0.0.5
mlflow
fastapi 
"uvicorn[standard]" 
python-multipart 
ultralytics 
Pillow